Initial draft of rtorrent XMLRPC API wrappers
This crate attempts to expose the rtorrent xmlrpc remote control
interface via a typed, Rusty interface.  For now, only a subset of
methods are exposed.

The top-level structure is `crate::Server`, which represents a logical
XMLRPC endpoint.  (For now, only HTTP endpoints are supported, but that
could be expanded relatively easily.)

One can get a list of loaded torrents via `Server::download_list()`.
`crate::Download` objects represent a loaded torrent (identified by SHA1
digest, in hex).

For each `Download`, there are a number of accessors for attributes on
that loaded torrent.  Accessors on `Download` correspond to the `d.*`
methods in the rtorrent API.  Additionally, one can get a list of
trackers for that download with `Download::trackes()`.

`crate::DownloadTracker` objects represent a specific tracker index for
a given download.  Accessors on `DownloadTracker` correspond to the
`t.*` methods in the rtorrent API.  One example is
`DownloadTracker::url()`.

Significant TODO:
* Write a README
* Document everything (somewhat limited by lack of upstream
documentation)
* Examples
* Publish to crates.io