Add File object and method to fetch files for a download
We can get `crate::File`s for a loaded torrent via the
`Download::files()` method.  Files represent the content associated with
a torrent.  There can be one or more files associated with any given
torrent.  Accessors on `File` correspond to the `f.*` methods in the
rtorrent API.  One example is `File::path()`.